Cleanup strata-client bin: get rid of l1_reader mod
