+++ crates/state/Cargo.toml
@@ -22,7 +22,7 @@ ethnum = "1.5.0"
 hex.workspace = true
 serde.workspace = true
 sha2.workspace = true
-tracing.workspace = true   # ideally this shouldn't be in this crate
+tracing.workspace = true     # ideally this shouldn't be in this crate
 
 
 [dev-dependencies]
